Refactor IndexLimitException to handle general db limits
