+++ src/Database/Database.php
@@ -6,7 +6,7 @@ use Exception;
 use Utopia\Database\Validator\Authorization;
 use Utopia\Database\Validator\Structure;
 use Utopia\Database\Exception\Authorization as AuthorizationException;
-use Utopia\Database\Exception\IndexLimit as IndexLimitException;
+use Utopia\Database\Exception\Limit as LimitException;
 use Utopia\Database\Exception\Structure as StructureException;
 use Utopia\Cache\Cache;
 
@@ -529,7 +529,7 @@ class Database
         $collection = $this->getCollection($collection);
 
         if ($this->adapter->getIndexCount($collection) >= $this->adapter->getIndexLimit()) {
-            throw new IndexLimitException('Index limit reached. Cannot create new index.');
+            throw new LimitException('Index limit reached. Cannot create new index.');
         }
 
         $collection->setAttribute('indexes', new Document([
@@ -621,7 +621,7 @@ class Database
         $collection = $this->getCollection($collection);
 
         if ($this->adapter->getIndexCount($collection) >= $this->adapter->getIndexLimit()) {
-            throw new IndexLimitException('Index limit reached. Cannot create new index.');
+            throw new LimitException('Index limit reached. Cannot create new index.');
         }
 
         $collection->setAttribute('indexesInQueue', new Document([
