+++ src/Database/Exception/Limit.php
